tp1/requirements.txt
psycopg2-binary>=2.9
pandas>=2.2
python-dateutil>=2.9

